lint and format
