+++ src/components/Root.jsx
@@ -1,4 +1,4 @@
-import { Link, Outlet } from 'react-router-dom'
+import { Outlet } from 'react-router-dom'
 import './Root.css'
 import { Navbar } from './Navbar'
 import { Footer } from './Footer'
