--- src/pages/Florida.jsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export default function Florida() {
-    return <div>Florida</div>
-}
